# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "FileMetadataDetail": {
        "properties": {
          "bitrate": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bitrate"
          },
          "codec": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codec"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "exif": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exif"
          },
          "extension": {
            "title": "Extension",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "image_height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Height"
          },
          "image_width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Width"
          },
          "md5": {
            "title": "Md5",
            "type": "string"
          },
          "metadata_warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Warning"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "pdf_author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Author"
          },
          "pdf_pages": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Pages"
          },
          "pdf_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Title"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "size_human": {
            "title": "Size Human",
            "type": "string"
          },
          "uploaded_at": {
            "format": "date-time",
            "title": "Uploaded At",
            "type": "string"
          }
        },
        "required": [
          "filename",
          "size_bytes",
          "size_human",
          "mime_type",
          "extension",
          "md5",
          "sha256",
          "uploaded_at"
        ],
        "title": "FileMetadataDetail",
        "type": "object"
      },
      "FileUploadResponse": {
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileMetadataDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "size_human": {
            "title": "Size Human",
            "type": "string"
          },
          "uploaded_at": {
            "format": "date-time",
            "title": "Uploaded At",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "key",
          "filename",
          "size_bytes",
          "size_human",
          "content_type",
          "uploaded_at"
        ],
        "title": "FileUploadResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PresignUploadRequest": {
        "description": "What the browser declares before uploading directly to B2.",
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "filename",
          "content_type",
          "size_bytes"
        ],
        "title": "PresignUploadRequest",
        "type": "object"
      },
      "PresignUploadResponse": {
        "description": "A short-lived presigned PUT the browser uploads to, plus the exact\nheaders it must send. `Content-Length` and `content-type` are signed into\nthe URL, so B2 rejects a body of any other size or type.",
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Headers",
            "type": "object"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "key",
          "url",
          "method",
          "content_type",
          "headers",
          "expires_in"
        ],
        "title": "PresignUploadResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerifyUploadRequest": {
        "description": "Sent after the direct PUT so the API can inspect the stored object.",
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          }
        },
        "required": [
          "key"
        ],
        "title": "VerifyUploadRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local API for the Vibe Coding Starter Kit template, providing file upload and management backed by Backblaze B2. This contract documents the template's local API, not a hosted public endpoint.",
    "title": "Vibe Coding Starter Kit API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/upload/presign": {
      "post": {
        "description": "Validate a declared upload and hand back a presigned PUT.\n\nThe browser uploads the bytes straight to B2 with the returned URL, so they\nnever traverse this Function \u2014 that is what lifts Vercel's ~4.5 MB payload\nceiling. Size and content-type are signed into the URL (see the service).",
        "operationId": "presign_upload_upload_presign_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Presign Upload",
        "tags": [
          "upload"
        ]
      }
    },
    "/upload/verify": {
      "post": {
        "description": "Confirm an object just uploaded directly to B2 is valid and visible.",
        "operationId": "verify_upload_route_upload_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Upload Route",
        "tags": [
          "upload"
        ]
      }
    }
  }
}